store-v1

+++ java-backend/src/main/java/com/fashion/modules/product/service/impl/ProductServiceImpl.java
@@ -12,16 +12,25 @@ import java.util.Map;
 import java.util.Set;
 import java.util.stream.Collectors;
 
+import javax.persistence.EntityManager;
 import javax.transaction.Transactional;
 
 import org.apache.commons.collections.CollectionUtils;
+import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.lang3.tuple.Pair;
+import org.apache.lucene.search.Query;
 import org.apache.poi.ss.usermodel.Row;
 import org.apache.poi.ss.usermodel.Workbook;
 import org.apache.poi.ss.usermodel.WorkbookFactory;
+import org.hibernate.search.jpa.FullTextEntityManager;
+import org.hibernate.search.jpa.Search;
+import org.hibernate.search.query.dsl.QueryBuilder;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.cache.annotation.CacheEvict;
+import org.springframework.cache.annotation.CachePut;
+import org.springframework.cache.annotation.Cacheable;
 import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.http.HttpEntity;
 import org.springframework.http.HttpHeaders;
@@ -90,10 +99,13 @@ public class ProductServiceImpl extends BaseService implements ProductService {
 
 	@Autowired
 	private ProductImageRepository imageRepo;
-	
+
 	@Autowired
 	private RestTemplate restTemplate;
-	
+
+	@Autowired
+	private EntityManager em;
+
 	private static final String PRODUCT_NAME = "Product Name";
 	private static final String PRICE = "Price";
 	private static final String COLOR = "Color";
@@ -108,12 +120,13 @@ public class ProductServiceImpl extends BaseService implements ProductService {
 
 	@Override
 	@Transactional
+	@CacheEvict(value = Constants.PRODUCTS, allEntries = true)
 	public ProductVM createProduct(final ProductReq req) {
 		final Store store = getStore(getUserContext());
 		final Integer storeId = store.getId();
 		final Category category = categoryRepo.findOneByIdAndStoreId(req.getCategoryId(), storeId);
 		if (category == null) {
-			throw new InvalidArgumentException(" Can't found category ");
+			throw new InvalidArgumentException(ErrorMessage.NOT_FOUND_CATEGORY);
 		}
 		final Product product = new Product();
 		product.setPrice(req.getPrice());
@@ -121,33 +134,27 @@ public class ProductServiceImpl extends BaseService implements ProductService {
 		product.setProductName(req.getProductName());
 		final Brand brand = brandRepo.findOneByIdAndStoreId(req.getBrandId(), storeId);
 		if (brand == null) {
-			throw new InvalidArgumentException(" Can't found brand ");
+			throw new InvalidArgumentException(ErrorMessage.NOT_FOUND_BRAND);
 		}
 		product.setBrand(brand);
 		product.setCategory(category);
 		final List<String> productImages = req.getImages();
 		if (CollectionUtils.isNotEmpty(productImages)) {
-			product.setProductImages(req.getImages().parallelStream().map(it -> {
-				return new ProductImage(it, product);
-			}).collect(Collectors.toSet()));
+			product.setProductImages(appendProductImages(req, product));
 		}
 		final Set<ProductDetailVM> productDetails = req.getProductDetails();
 
 		if (CollectionUtils.isNotEmpty(productDetails)) {
-			product.setProductDetails(productDetails.parallelStream().map(it -> {
-				final Size size = sizeRepo.findOneByIdAndStoreId(it.getSizeId(), storeId);
-				final Color color = colorRepo.findOneByIdAndStore(it.getColorId(), storeId);
-				if (size == null) {
-					throw new InvalidArgumentException(" Size doesn't exit ");
-				}
-				if (color == null) {
-					throw new InvalidArgumentException(" Color doesn't exit ");
-				}
-				return new ProductDetail(it.getQuantity(), product, size, color);
-			}).collect(Collectors.toSet()));
+			product.setProductDetails(appendProductDetails(storeId, product, productDetails));
 		}
 		productRepo.save(product);
 		return convertProductToVM(product);
+	}
+
+	private Set<ProductImage> appendProductImages(final ProductReq req, final Product product) {
+		return req.getImages().parallelStream().map(it -> {
+			return new ProductImage(it, product);
+		}).collect(Collectors.toSet());
 	}
 
 	private ProductVM convertProductToVM(final Product product) {
@@ -167,8 +174,8 @@ public class ProductServiceImpl extends BaseService implements ProductService {
 		vm.setProductDetails(product.getProductDetails().stream().map(it -> {
 			final Color color = it.getColor();
 			final Size size = it.getSize();
-			return new ProductDetailVM(size.getId(), size.getSizeName(), color.getId(), color.getColorName(),
-					color.getcolorCode(), it.getQuantity());
+			return new ProductDetailVM(it.getId(), size.getId(), size.getSizeName(), color.getId(),
+					color.getColorName(), color.getcolorCode(), it.getQuantity());
 		}).collect(Collectors.toSet()));
 		vm.setComments(product.getComments().stream().filter(i -> i.getEmail() != null)
 				.sorted(Comparator.comparing(Comment::getCreatedAt).reversed()).map(it -> convertToVM(it))
@@ -187,12 +194,13 @@ public class ProductServiceImpl extends BaseService implements ProductService {
 
 	@Override
 	@Transactional
+	@CachePut(value = Constants.PRODUCTS)
 	public ProductVM updateProduct(final ProductReq req) {
 		final Store store = getStore(getUserContext());
 		final Integer storeId = store.getId();
 		final Product product = productRepo.findOneProductByIdAndStore(req.getId(), storeId);
 		if (product == null) {
-			throw new NullPointerException("Not found product ");
+			throw new NullPointerException(ErrorMessage.NOT_FOUND);
 		}
 		final Category category = categoryRepo.findOneByIdAndStoreId(req.getCategoryId(), storeId);
 		product.setPrice(req.getPrice());
@@ -200,23 +208,11 @@ public class ProductServiceImpl extends BaseService implements ProductService {
 		product.setProductName(req.getProductName());
 		final List<String> productImages = req.getImages();
 		if (CollectionUtils.isNotEmpty(productImages)) {
-			product.setProductImages(req.getImages().parallelStream().map(it -> {
-				return new ProductImage(it, product);
-			}).collect(Collectors.toSet()));
+			product.setProductImages(appendProductImages(req, product));
 		}
 		final Set<ProductDetailVM> productDetails = req.getProductDetails();
 		if (CollectionUtils.isNotEmpty(productDetails)) {
-			product.setProductDetails(productDetails.parallelStream().map(it -> {
-				final Size size = sizeRepo.findOneByIdAndStoreId(it.getSizeId(), storeId);
-				final Color color = colorRepo.findOneByIdAndStore(it.getColorId(), storeId);
-				if (size == null) {
-					throw new InvalidArgumentException(" Size doesn't exit ");
-				}
-				if (color == null) {
-					throw new InvalidArgumentException(" Color doesn't exit ");
-				}
-				return new ProductDetail(it.getQuantity(), product, size, color);
-			}).collect(Collectors.toSet()));
+			product.setProductDetails(appendProductDetails(storeId, product, productDetails));
 		}
 		product.setBrand(brandRepo.findOneByIdAndStoreId(req.getBrandId(), storeId));
 		product.setCategory(category);
@@ -224,12 +220,29 @@ public class ProductServiceImpl extends BaseService implements ProductService {
 		return convertProductToVM(product);
 	}
 
+	private Set<ProductDetail> appendProductDetails(final Integer storeId, final Product product,
+			final Set<ProductDetailVM> productDetails) {
+		return productDetails.parallelStream().map(it -> {
+			final Size size = sizeRepo.findOneByIdAndStoreId(it.getSizeId(), storeId);
+			final Color color = colorRepo.findOneByIdAndStore(it.getColorId(), storeId);
+			if (size == null) {
+				throw new InvalidArgumentException(ErrorMessage.NOT_FOUND_SIZE);
+			}
+			if (color == null) {
+				throw new InvalidArgumentException(ErrorMessage.NOT_FOUND_COLOR);
+			}
+			return new ProductDetail(it.getQuantity(), product, size, color);
+		}).collect(Collectors.toSet());
+	}
+
 	@Override
 	@Transactional
 	public Page<ProductVM> getAllProductByStore(final Integer page, final Integer pageSize, final ProductReq req) {
 		if (req == null) {
-			return productRepo.findAllProductStore(getCurrentStoreId(), PageRequest.of(page, pageSize))
-					.map(it -> convertProductToVM(it));
+			@SuppressWarnings("null")
+			final Integer storeId = req.getStoreId();
+			return productRepo.findAllProductStore(storeId != null ? storeId : getCurrentStoreId(),
+					PageRequest.of(page, pageSize)).map(it -> convertProductToVM(it));
 		}
 		return filterProduct(page, pageSize, req);
 	}
@@ -251,12 +264,15 @@ public class ProductServiceImpl extends BaseService implements ProductService {
 			res.setQuantity(it.getQuantity());
 			res.setCategoryName(product.getCategory().getCategoryName());
 			res.setColorHex(color.getcolorCode());
+			res.setStoreId(product.getStore().getId());
+			res.setImage(product.getProductImages().stream().map(i -> i.getImage()).findFirst().get());
 			return res;
 		}).collect(Collectors.toList());
 	}
 
 	@Override
 	@Transactional
+	@Cacheable(value = Constants.PRODUCTS)
 	public Page<ProductVM> searchProductByKeywordAndStore(final String keyword, final Integer page,
 			final Integer pageSize) {
 		return productRepo.searchByKeywordAndStore(keyword, getCurrentStoreId(), PageRequest.of(page, pageSize))
@@ -266,14 +282,22 @@ public class ProductServiceImpl extends BaseService implements ProductService {
 	@Override
 	@Transactional
 	public Page<ProductVM> filterProduct(final Integer page, final Integer pageSize, final ProductReq req) {
-		return productRepo.filterProduct(page, pageSize, getCurrentStoreId(), req).map(it -> convertProductToVM(it));
+		final Integer storeId = req.getStoreId();
+		return productRepo
+				.filterProduct(page, pageSize, storeId != null ? storeId : getCurrentStoreId(), req)
+				.map(it -> convertProductToVM(it));
 	}
 
 	@Override
 	@Transactional
+	@CacheEvict(value = Constants.PRODUCTS, allEntries = true)
 	public ProductVM deleteProduct(final Integer id, final Integer page, final Integer pageSize,
 			final SortType sortOrder, final String sortField) {
 		try {
+			final Product product = productRepo.findOneProductByIdAndStore(id, getCurrentStoreId());
+			if (product == null) {
+				throw new InvalidArgumentException(ErrorMessage.NOT_FOUND);
+			}
 			productRepo.deleteById(id);
 			final Page<ProductVM> products = filterProduct(page, pageSize, new ProductReq(sortOrder, sortField));
 			final List<ProductVM> content = products.getContent();
@@ -282,10 +306,7 @@ public class ProductServiceImpl extends BaseService implements ProductService {
 			}
 			return Iterables.getLast(content);
 		} catch (Exception e) {
-			if (e instanceof DataIntegrityViolationException) {
-				throw new InvalidArgumentException(" Product existed in carts. You can't delete this.");
-			}
-			throw new InvalidArgumentException("Product not found ");
+			throw new InvalidArgumentException(ErrorMessage.PRODUCT_EXISTED_CART);
 
 		}
 	}
@@ -344,7 +365,7 @@ public class ProductServiceImpl extends BaseService implements ProductService {
 			}
 		});
 		if (headerTexts.size() != map.keySet().size()) {
-			throw new InvalidArgumentException("Please take a check header row.");
+			throw new InvalidArgumentException(ErrorMessage.HEADER_ERROR);
 		}
 		return map;
 	}
@@ -355,6 +376,7 @@ public class ProductServiceImpl extends BaseService implements ProductService {
 
 	@Override
 	@Transactional
+	@Cacheable(Constants.PRODUCTS)
 	public List<ProductVM> createProducts(final List<ProductRes> products) {
 		final Integer storeId = getCurrentStoreId();
 		final PageRequest page = PageRequest.of(0, Integer.MAX_VALUE);
@@ -412,6 +434,7 @@ public class ProductServiceImpl extends BaseService implements ProductService {
 
 	@Override
 	@Transactional
+	@Cacheable(Constants.PRODUCTS)
 	public Page<ProductVM> getAllOrFilterProduct(final ProductFilterRequest req, final Integer page,
 			final Integer pageSize) {
 		return productRepo.filterProduct(req, page, pageSize).map(it -> convertProductToVM(it));
@@ -419,13 +442,14 @@ public class ProductServiceImpl extends BaseService implements ProductService {
 
 	@Override
 	@Transactional
-	public Page<ProductVM> getBestSellerProductByStore(final Integer storeId, final Integer page,
-			final Integer pageSize) {
+	@Cacheable(Constants.BEST_SELLER)
+	public List<ProductVM> getBestSellerProductByStore(final Integer storeId) {
 		final BestSellerData data = getBestSellerByStore(storeId);
 		return productRepo
-				.getBestSeller(storeId, page, pageSize, data.getProductDetails().stream()
-						.map(it -> it.getProductDetailId()).collect(Collectors.toSet()))
-				.map(it -> convertProductToVM(it));
+				.getBestSeller(storeId,
+						data.getProductDetails().stream().map(it -> it.getProductDetailId())
+								.collect(Collectors.toSet()))
+				.stream().map(i -> convertProductToVM(i)).collect(Collectors.toList());
 	}
 
 	private BestSellerData getBestSellerByStore(final Integer storeId) {
@@ -436,6 +460,58 @@ public class ProductServiceImpl extends BaseService implements ProductService {
 		}
 		return restTemplate.exchange(builder.build().encode().toUri(), HttpMethod.GET,
 				new HttpEntity<String>(new HttpHeaders()), BestSellerRes.class).getBody().getData();
+	}
+
+	@SuppressWarnings("unchecked")
+	@Override
+	@Transactional
+	public Page<ProductVM> searchProductByKeyword(final Integer page, final Integer pageSize, final String keyword) {
+		if (StringUtils.isNotEmpty(keyword)) {
+			final FullTextEntityManager fullTextEntityManager = Search.getFullTextEntityManager(em);
+			final QueryBuilder qb = fullTextEntityManager.getSearchFactory().buildQueryBuilder()
+					.forEntity(Product.class).get();
+			final Query productQuery = qb.keyword().onFields("productName", "price", "brand.brandName",
+					"productDetails.color.colorName", "category.categoryName", "productDetails.size.sizeName")
+					.matching(keyword).createQuery();
+			final org.hibernate.search.jpa.FullTextQuery fullTextQuery = fullTextEntityManager
+					.createFullTextQuery(productQuery, Product.class);
+			final List<Product> products = fullTextQuery.getResultList();
+			final Map<Integer, String> brandIdMap = brandRepo
+					.findBrandByIdIn(
+							products.parallelStream().map(it -> it.getBrand().getId()).collect(Collectors.toSet()))
+					.parallelStream().collect(Collectors.toMap(Brand::getId, Brand::getBrandName));
+			final Map<Integer, String> categoryIdMap = categoryRepo
+					.findCategoryByIdIn(
+							products.parallelStream().map(it -> it.getCategory().getId()).collect(Collectors.toSet()))
+					.parallelStream().collect(Collectors.toMap(Category::getId, Category::getCategoryName));
+//			final Map<Product, List<ProductImage>> productImageMap = imageRepo
+//					.getProductImageByIdIn(
+//							products.parallelStream().flatMap(it -> it.getProductImages().stream().map(i -> i.getId()))
+//									.collect(Collectors.toList()))
+//					.parallelStream().collect(Collectors.groupingBy(ProductImage::getProduct, Collectors.toList()));
+			fullTextQuery.setMaxResults(pageSize);
+			fullTextQuery.setFirstResult(page * pageSize);
+
+			final List<Product> products2 = fullTextQuery.getResultList();
+			return new PageImpl<ProductVM>(products2.stream().map(it -> {
+				final ProductVM vm = new ProductVM();
+				vm.setId(it.getId());
+				vm.setPrice(it.getPrice());
+				vm.setProductName(it.getProductName());
+				final Integer brandId = it.getBrand().getId();
+				final Integer categoryId = it.getCategory().getId();
+				vm.setBrandName(brandIdMap.get(brandId));
+				vm.setBrandId(brandId);
+				vm.setCategoryId(categoryId);
+				vm.setCategoryName(categoryIdMap.get(categoryId));
+				vm.setStoreId(it.getStore().getId());
+				vm.setProductImages(it.getProductImages().stream().map(i -> mapper.map(i, ProductImageVM.class))
+						.collect(Collectors.toSet()));
+				return vm;
+			}).collect(Collectors.toList()), PageRequest.of(page, pageSize), products.size());
+		}
+		return new PageImpl<ProductVM>(Collections.emptyList());
+
 	}
 
 }
